Rename class Position to PhysicalPosition and add onNewPhysicalPosition to EventHandler
The purpose of renaming the class is to differentiate between a move
(leading to a new "chess position") and a change on the physical board that does not
necessarily equate to a legal move.

The new event (onNewPhysicalPosition) allows for a "diagnostics" view in the Activity,
displaying the current board data.